♻️ refactor:  enumをエンティティに切り出し
Issue #145
