+++ lib/domain/entities/repo_search_repos_order.dart
@@ -1,0 +1,16 @@
+// Copyright 2022 susatthi All rights reserved.
+// Use of this source code is governed by a MIT license that can be
+// found in the LICENSE file.
+
+/// リポジトリ検索用オーダー
+enum RepoSearchReposOrder {
+  desc,
+  asc;
+
+  /// 文字列からオーダーを返す
+  /// 見つからない場合は IterableElementError.noElement() を投げる
+  static RepoSearchReposOrder valueOf(String name) {
+    return RepoSearchReposOrder.values
+        .firstWhere((element) => element.name == name);
+  }
+}
