+++ lib/infrastructure/github/repo_repository.dart
@@ -5,6 +5,8 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../domain/entities/repo.dart';
+import '../../domain/entities/repo_search_repos_order.dart';
+import '../../domain/entities/repo_search_repos_sort.dart';
 import '../../domain/entities/search_repos_result.dart';
 import '../../domain/repositories/repo_repository.dart';
 import 'api.dart';
