+++ test/domain/repositories/repo_repository_test.dart
@@ -4,6 +4,8 @@
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:github_search/domain/entities/repo_search_repos_order.dart';
+import 'package:github_search/domain/entities/repo_search_repos_sort.dart';
 import 'package:github_search/domain/repositories/repo_repository.dart';
 
 import '../../test_utils/locale.dart';
