+++ test/presentation/repo/components/search_repos_sort_test.dart
@@ -6,7 +6,7 @@ import 'dart:io';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:github_search/domain/repositories/repo_repository.dart';
+import 'package:github_search/domain/entities/repo_search_repos_sort.dart';
 import 'package:github_search/presentation/repo/components/search_repos_sort.dart';
 
 import '../../../test_utils/hive.dart';
